The Cluster Buster
The Cluster Buster is a project undertaken to develop a device which will help victims of cluster headaches and allow the user more information and control on their condition.
The Cluster Buster is a non-invasive vagus nerve stimulating device that will allow the user self-administer massage therapy without inserting a medical device or instrument into the body. I explored various sensors, software, and programming languages to develop the cluster Buster as an all-in-one handheld device to improve quality of life for victims. This device will hopefully be an alternative form of treatment rather than invasive injections or lifelong medicine.